+++ examples/generate_error_report.rs
@@ -55,6 +55,10 @@ where
     TetrahedronMagnet<T>: Source<T>,
     TriangleMagnet<T>: Source<T>,
     MeshMagnet<T>: Source<T>,
+    CircularCurrent<T>: Source<T>,
+    PathCurrent<T>: Source<T>,
+    SheetCurrent<T>: Source<T>,
+    TriangleCurrent<T>: Source<T>,
 {
     let f = |v: f64| T::from(v).unwrap();
 
@@ -71,6 +75,40 @@ where
     ];
 
     match name {
+        "CircularCurrent" => Box::new(CircularCurrent::new(pos, rot, f(1.0), f(1.0))),
+        "PathCurrent" => Box::new(PathCurrent::new(
+            pos,
+            rot,
+            f(100.0),
+            vec![
+                vector![f(-0.1), f(-0.1), f(-0.1)],
+                vector![f(0.1), f(-0.1), f(-0.1)],
+                vector![f(0.0), f(0.1), f(-0.1)],
+                vector![f(0.0), f(0.0), f(0.1)],
+            ],
+        )),
+        "SheetCurrent" => {
+            let mesh = TriMesh::<T>::new_unchecked(
+                tetrahedron_vertices,
+                vec![[0, 2, 1], [0, 1, 3], [1, 2, 3], [0, 3, 2]],
+            );
+            Box::new(SheetCurrent::new(
+                pos,
+                rot,
+                vec![vector![f(1.0), f(2.0), f(3.0)]; 4],
+                mesh,
+            ))
+        }
+        "TriangleCurrent" => Box::new(TriangleCurrent::new(
+            pos,
+            rot,
+            vector![f(1.0), f(2.0), f(3.0)],
+            [
+                vector![f(-0.1), f(-0.1), f(-0.1)],
+                vector![f(0.1), f(-0.1), f(-0.1)],
+                vector![f(0.0), f(0.1), f(-0.1)],
+            ],
+        )),
         "CylinderMagnet" => Box::new(CylinderMagnet::new(pos, rot, pol, f(0.1), f(0.2))),
         "CuboidMagnet" => Box::new(CuboidMagnet::new(pos, rot, pol, [f(0.1), f(0.2), f(0.3)])),
         "Dipole" => Box::new(Dipole::new(pos, rot, pol)),
@@ -114,14 +152,17 @@ fn add_magnet_accuracy<T: magba::base::Float + std::str::FromStr>(
 
     Ok(report.with_row(
         Row::new_batch(name, source.as_batch_evaluator())
-            .with_split_csv_cases(points_path, ref_path)?,
+            .with_split_csv_cases(points_path, ref_path)?
+            .with_criterion_id(format!("fields_{}/{}", std::any::type_name::<T>(), name)),
     ))
 }
 
 fn main() -> Result<(), Box<dyn Error>> {
-    println!("# Magba Accuracy Report\n");
-
     let magnets = [
+        ("CircularCurrent", "circularcurrent.csv"),
+        ("PathCurrent", "polyline.csv"),
+        ("SheetCurrent", "sheetcurrent.csv"),
+        ("TriangleCurrent", "trianglecurrent.csv"),
         ("CylinderMagnet", "cylinder.csv"),
         ("CuboidMagnet", "cuboid.csv"),
         ("Dipole", "dipole.csv"),
@@ -139,31 +180,37 @@ fn main() -> Result<(), Box<dyn Error>> {
             .with_column(Column::accuracy("Mean").with_stat(ColumnStat::Mean))
             .with_column(Column::accuracy("P95").with_stat(ColumnStat::P95))
             .with_column(Column::accuracy("Max").with_stat(ColumnStat::Max))
+            .with_column(Column::<f64>::perf("Performance").postprocess(|perf| perf / 1000.0)) // Divide by the number of test points
     }
+
+    let mut content = std::fs::read_to_string("tests/report_template.md")
+        .expect("Cannot read the report template");
+
+    content = content.replace("{{TEST_ENV}}", &current_env!().to_string());
 
     // f64 report
     {
-        println!("## Relative Error: f64\n");
         let mut report = get_report();
 
         for &(name, ref_file) in &magnets {
             let magnet = create_magnet::<f64>(name);
             report = add_magnet_accuracy(report, name, magnet, ref_file)?;
         }
-        println!("{}", report.render_markdown());
+        content = content.replace("{{ERROR_F64}}", report.render_markdown().as_str());
     }
 
     // f32 report
     {
-        println!("\n## Relative Error: f32\n");
         let mut report = get_report();
 
         for &(name, ref_file) in &magnets {
             let magnet = create_magnet::<f32>(name);
             report = add_magnet_accuracy(report, name, magnet, ref_file)?;
         }
-        println!("{}", report.render_markdown());
+        content = content.replace("{{ERROR_F32}}", report.render_markdown().as_str());
     }
 
+    let _ = std::fs::write("tests/README.md", content).expect("Cannot write the report");
+    println!("tests/README.md generated successfully");
     Ok(())
 }